requirements.txt
Django==1.5.1
python-openid>=2.0.0
oauth>=1.0.0
django_compressor>=1.3
South>=0.8
django-admin-tools>=0.5
pytz>=2013b
django-crispy-forms>=1.3.2
django-dajaxice==0.5.5

Pillow

oauth2>=1.5.211
django-social-auth>=0.7.22

git+http://github.com/pythonforfacebook/facebook-sdk.git#egg=facebook_sdk

easy_thumbnails

django-countries
django-tinymce
